Re-brand isingmodel project to spyns * Rename isingmodel directory to spyns * Refactor code to use spyns project directory * Refactor tests to use spyns project directory * Refactor docs files to use spyns project directory and new name * Refactor build and setup files to use spyns directory and new name * Updates to dev environment and files   * Pin dev environment to Anaconda   * Use conventional yml extension on conda environment file
