+++ docs/index.rst
@@ -1,5 +1,5 @@
-isingmodel documentation
-========================
+sPyns documentation
+===================
 
 .. include:: ../README.rst
    :start-line: 2
@@ -11,11 +11,11 @@ API
 .. autosummary::
    :toctree: _autosummary
 
-   isingmodel
-   isingmodel.algorithms
-   isingmodel.data
-   isingmodel.distributions
-   isingmodel.model
-   isingmodel.run
-   isingmodel.sampling
-   isingmodel.statistics
+   spyns
+   spyns.algorithms
+   spyns.data
+   spyns.distributions
+   spyns.model
+   spyns.run
+   spyns.sampling
+   spyns.statistics
