--- environment.yaml
@@ -1,33 +1,0 @@
----
-name: ising-model-py
-channels:
-  - defaults
-  - jkglasbrenner
-dependencies:
-  - attrs
-  - autopep8
-  - click
-  - cython
-  - flake8
-  - ipython
-  - jedi
-  - jupyter
-  - numpy
-  - matplotlib
-  - pandas
-  - pyhamcrest
-  - pylint
-  - pytest
-  - pytest-runner
-  - python=3.7.0
-  - rope
-  - ruamel_yaml
-  - scipy
-  - sphinx
-  - sphinx_rtd_theme
-  - yapf
-  - pip:
-      - pip==18.1
-      - pydocstyle
-      - sphinx-autodoc-typehints
-...
