+++ environment.yml
@@ -1,0 +1,18 @@
+---
+name: spyns
+channels:
+  - defaults
+  - jkglasbrenner
+dependencies:
+  - anaconda=5.3.0=py37_0
+  - autopep8
+  - pyhamcrest
+  - pytest-runner
+  - sphinx_rtd_theme
+  - yapf
+  - pip:
+    - flake8
+    - pip==18.1
+    - pydocstyle
+    - sphinx-autodoc-typehints
+...
