--- isingmodel/__init__.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import isingmodel.algorithms
-import isingmodel.data
-import isingmodel.lattice
-import isingmodel.model
-import isingmodel.run
-import isingmodel.sampling
-import isingmodel.statistics
