+++ spyns/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+import spyns.algorithms
+import spyns.data
+import spyns.lattice
+import spyns.model
+import spyns.run
+import spyns.sampling
+import spyns.statistics
